rename util and setup drop statement for migration
