+++ packages/server/src/main.ts
@@ -6,10 +6,10 @@ import { AppModule } from './app.module'
 import { Logger, ValidationPipe } from '@nestjs/common'
 import { isCli, isProd } from '../utils/environment.util'
 import { PROCESS_NAME } from '../utils/constants'
-import { updateSettings } from '../utils/wakatime.util'
+import { updateSettings } from '../utils/ini.util'
 import { startMigrations } from './v1/database/migrations'
 
-const updatedIniValues: Record<string, string> = {
+const updatedWakatimeIniValues: Record<string, string> = {
   api_key: 'eacb3beb-dad8-4fa1-b6ba-f89de8bf8f4a', // placeholder value
   api_url: 'http://localhost:14400/api/v1/wakatime',
 }
@@ -38,7 +38,7 @@ export async function bootstrap() {
     }),
   )
 
-  await updateSettings(updatedIniValues)
+  await updateSettings(updatedWakatimeIniValues)
   await startMigrations()
   await app.listen(port)
   process.title = PROCESS_NAME
